tl;dr New build system! Use "grunt dev" for a local dev server. It automagically minifes everything as it does.
Create a Gruntfile with
- uglify (to minify js)
- cssmin (to minify css)
- imagemin (" images)
- copy (to move html + fonts)

Moved all development assets to a new assets folder.
All work should be done in there.

The static folder now holds built files.

Grunt newer is used to ensure that we only build the extra files we need to every time we build.

The default task starts a production server, and the dev task starts a development server.

Removed the app-local.js file - it's no longer needed.